+++ src/com/martin/ELib/Application.java
@@ -1,4 +1,4 @@
-package com.martin.eDocManager;
+package com.martin.ELib;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
