+++ src/com/martin/ELib/annotation/MethodDetails.java
@@ -1,4 +1,4 @@
-package com.martin.eDocManager.annotation;
+package com.martin.ELib.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
